@@ -4,13 +4,12 @@ Render to qtquick from agg.
 
 import numpy as np
 from matplotlib.backends.backend_agg import FigureCanvasAgg
-from .backend_qtquick import (
-    QtCore, QtGui, FigureCanvasQtQuick)
+from .backend_qtquick import QtCore, QtGui, FigureCanvasQtQuick
 
 
 class FigureCanvasQtQuickAgg(FigureCanvasAgg, FigureCanvasQtQuick):
-    """ This class customizes the FigureCanvasQtQuick for Agg
-    """
+    """This class customizes the FigureCanvasQtQuick for Agg"""
+
     def __init__(self, figure=None, parent=None):
         super().__init__(figure=figure, parent=parent)
         self.blitbox = None
@@ -25,7 +24,7 @@ class FigureCanvasQtQuickAgg(FigureCanvasAgg, FigureCanvasQtQuick):
 
         # if the canvas does not have a renderer, then give up and wait for
         # FigureCanvasAgg.draw(self) to be called
-        if not hasattr(self, 'renderer'):
+        if not hasattr(self, "renderer"):
             return
 
         if self.blitbox is None:
@@ -46,10 +45,13 @@ class FigureCanvasQtQuickAgg(FigureCanvasAgg, FigureCanvasQtQuick):
                 stringBuffer = self.renderer.tostring_argb()
 
             # convert the Agg rendered image -> qImage
-            qImage = QtGui.QImage(stringBuffer, int(self.renderer.width),
-                                  int(self.renderer.height),
-                                  QtGui.QImage.Format_RGBA8888)
-            if hasattr(qImage, 'setDevicePixelRatio'):
+            qImage = QtGui.QImage(
+                stringBuffer,
+                int(self.renderer.width),
+                int(self.renderer.height),
+                QtGui.QImage.Format_RGBA8888,
+            )
+            if hasattr(qImage, "setDevicePixelRatio"):
                 # Not available on Qt4 or some older Qt5.
                 qImage.setDevicePixelRatio(self.dpi_ratio)
             # get the rectangle for the image
@@ -70,15 +72,14 @@ class FigureCanvasQtQuickAgg(FigureCanvasAgg, FigureCanvasQtQuick):
             t = b + h
             reg = self.copy_from_bbox(bbox)
             stringBuffer = reg.to_string_argb()
-            qImage = QtGui.QImage(stringBuffer, w, h,
-                                  QtGui.QImage.Format_RGBA8888)
+            qImage = QtGui.QImage(stringBuffer, w, h, QtGui.QImage.Format_RGBA8888)
 
-            if hasattr(qImage, 'setDevicePixelRatio'):
+            if hasattr(qImage, "setDevicePixelRatio"):
                 # Not available on Qt4 or some older Qt5.
                 qImage.setDevicePixelRatio(self.dpi_ratio)
             pixmap = QtGui.QPixmap.fromImage(qImage)
 
-            p.drawPixmap(QtCore.QPoint(l, self.renderer.height-t), pixmap)
+            p.drawPixmap(QtCore.QPoint(l, self.renderer.height - t), pixmap)
 
             # draw the zoom rectangle to the QPainter
             self._draw_rect_callback(p)
@@ -98,7 +99,7 @@ class FigureCanvasQtQuickAgg(FigureCanvasAgg, FigureCanvasQtQuick):
         # repaint uses logical pixels, not physical pixels like the renderer.
         l, b, w, h = [pt / self._dpi_ratio for pt in bbox.bounds]
         t = b + h
-        self.repaint(l, self.renderer.height-t, w, h)
+        self.repaint(l, self.renderer.height - t, w, h)
 
     def print_figure(self, *args, **kwargs):
         super().print_figure(*args, **kwargs)
