@@ -34,6 +34,7 @@ import json
 import platform
 from .MatPlotLib import Canvas
 from .Model import Model
+from .DataHandler import DataHandler
 
 
 class SinglePlot(QObject):
@@ -44,10 +45,11 @@ class SinglePlot(QObject):
     fill_plot_page_signal = pyqtSignal(QJsonValue, arguments="props")
     plot_signal = pyqtSignal()
 
-    def __init__(self, canvas, model, messageHandler):
+    def __init__(self, canvas, model, datahandler, messageHandler):
         super().__init__()
         self.canvas: Canvas = canvas
         self.model: Model = model
+        self.datahandler: DataHandler = datahandler
         self.path = ""
         self.msg = messageHandler
 
@@ -96,13 +98,17 @@ class SinglePlot(QObject):
     @pyqtSlot(QJsonValue)
     def get_plot_data(self, plot_data):
         self.model.reset()
-        plot_data = plot_data.toVariant()
+        self.datahandler.reset()
+        plot_data: dict = plot_data.toVariant()
         canvas_props = plot_data["canvasProps"]
         dataProps = plot_data["dataProps"]
         fit_props = plot_data["fitProps"]
 
         # Loading data from the table
-        self.model.loadDataTable(plot_data["data"])
+        self.datahandler.loadDataTable(plot_data["data"])
+        self.model.data = self.datahandler.data
+        self.model._has_sx = self.datahandler.has_sx
+        self.model._has_sy = self.datahandler.has_sy
 
         # Getting function to fit
         # Anti-dummies system
@@ -122,9 +128,7 @@ class SinglePlot(QObject):
         # Setting expression
         if len(expIndVar) == 2:
             if self.model.exp_model != expIndVar[0]:
-                self.model.set_expression(
-                    expIndVar[0].strip(), expIndVar[1].strip()
-                )
+                self.model.set_expression(expIndVar[0].strip(), expIndVar[1].strip())
         elif len(expIndVar) == 1:
             if self.model.exp_model != expIndVar[0]:
                 self.model.set_expression(expIndVar[0])
@@ -194,9 +198,7 @@ class SinglePlot(QObject):
         self, x, y, sy, sx, kargs_errorbar, kargs_scatter, y_r=None, ssy=None
     ):
         """Macro para o plot dos dados."""
-        self.canvas.plot_error_bar_new(
-            x, y, sy, sx, kargs_errorbar, y_r=y_r, ssy=ssy
-        )
+        self.canvas.plot_error_bar_new(x, y, sy, sx, kargs_errorbar, y_r=y_r, ssy=ssy)
         self.canvas.plot_scatter(x, y, kargs_scatter, y_r)
 
     def plot(self, model: Model, canvas_props, fit_props, data_props):
@@ -240,7 +242,7 @@ class SinglePlot(QObject):
                 canvas_props["yaxis"].strip(),
                 partial_titles[1].strip(),
             ]
-        if model._has_data:
+        if self.datahandler._has_data:
             # Fitting expression to data, if there"s any expression
             if fit_props["adjust"]:
                 if model.exp_model != "":
@@ -274,9 +276,12 @@ class SinglePlot(QObject):
                 self.canvas.clear_axis()
                 # Getting data
                 x, y, sy, sx = model.data
-                inliers, outliers = model.inliers, model.outliers
+                _, outliers = model.inliers, model.outliers
                 c = np.array([list(colors.to_rgba(symbol_color))] * len(x))
-                c[outliers.astype(int), 3] = self.canvas.user_alpha_outliers
+                try:
+                    c[outliers.astype(int), 3] = self.canvas.user_alpha_outliers
+                except IndexError:
+                    pass
                 kargs_errorbar["ecolor"] = c
                 kargs_scatter["c"] = c
                 y_r = None
@@ -287,9 +292,7 @@ class SinglePlot(QObject):
                     # y_ri, y_ro = np.copy(y_r), np.array([])
                 if residuals:
                     self.canvas.switch_axes(hide_axes2=False)
-                    if (
-                        sigma_x and sigma_y
-                    ):  # Caso considerar as duas incertezas
+                    if sigma_x and sigma_y:  # Caso considerar as duas incertezas
                         ssy = model.predictInc(not sigma_x)
                         self.plot_data(
                             x,
@@ -305,9 +308,7 @@ class SinglePlot(QObject):
                         sigma_x is False and sigma_y is False
                     ):  # Caso desconsiderar as duas
                         self.canvas.plot_scatter(x, y, kargs_scatter, y_r)
-                    elif (
-                        sigma_x is False and sigma_y is True
-                    ):  # Caso considerar só sy
+                    elif sigma_x is False and sigma_y is True:  # Caso considerar só sy
                         ssy = model.predictInc(not sigma_x)
                         sx = np.array([0] * len(x))
                         self.plot_data(
@@ -418,44 +419,30 @@ class SinglePlot(QObject):
                             self.canvas.axes1.figure.canvas.draw_idle()
 
                     self.canvas.axes1.remove_callback(self.canvas.oid)
-                    self.canvas.axes1.figure.canvas.mpl_disconnect(
-                        self.canvas.cid
-                    )
+                    self.canvas.axes1.figure.canvas.mpl_disconnect(self.canvas.cid)
                     self.canvas.oid = self.canvas.axes1.callbacks.connect(
                         "xlim_changed", update
                     )
-                    self.canvas.cid = (
-                        self.canvas.axes1.figure.canvas.mpl_connect(
-                            "resize_event", update
-                        )
+                    self.canvas.cid = self.canvas.axes1.figure.canvas.mpl_connect(
+                        "resize_event", update
                     )
                 else:
                     self.canvas.clear_axis()
                     self.canvas.switch_axes(hide_axes2=True)
 
                     # Making Plots
-                    if (
-                        sigma_x and sigma_y
-                    ):  # Caso considerar as duas incertezas
-                        self.plot_data(
-                            x, y, sy, sx, kargs_errorbar, kargs_scatter
-                        )
+                    if sigma_x and sigma_y:  # Caso considerar as duas incertezas
+                        self.plot_data(x, y, sy, sx, kargs_errorbar, kargs_scatter)
                     elif (
                         sigma_x is False and sigma_y is False
                     ):  # Caso desconsiderar as duas
                         self.canvas.plot_scatter(x, y, kargs_scatter)
-                    elif (
-                        sigma_x is False and sigma_y is True
-                    ):  # Caso considerar só sy
+                    elif sigma_x is False and sigma_y is True:  # Caso considerar só sy
                         sx = np.array([0] * len(x))
-                        self.plot_data(
-                            x, y, sy, sx, kargs_errorbar, kargs_scatter
-                        )
+                        self.plot_data(x, y, sy, sx, kargs_errorbar, kargs_scatter)
                     else:  # Caso considerar só sx
                         sy = np.array([0] * len(x))
-                        self.plot_data(
-                            x, y, sy, sx, kargs_errorbar, kargs_scatter
-                        )
+                        self.plot_data(x, y, sy, sx, kargs_errorbar, kargs_scatter)
 
                     self.canvas.set_axes_props_without_axes_2(
                         xmin, xmax, xdiv, ymin, ymax, ydiv, grid, log_x, log_y
@@ -521,9 +508,7 @@ class SinglePlot(QObject):
                             self.canvas.axes1.figure.canvas.draw_idle()
 
                     self.canvas.axes1.remove_callback(self.canvas.oid)
-                    self.canvas.axes1.figure.canvas.mpl_disconnect(
-                        self.canvas.cid
-                    )
+                    self.canvas.axes1.figure.canvas.mpl_disconnect(self.canvas.cid)
                     self.canvas.oid = self.canvas.axes1.callbacks.connect(
                         "xlim_changed", update
                     )
@@ -535,7 +520,7 @@ class SinglePlot(QObject):
                 self.canvas.clear_axis()
                 self.canvas.switch_axes(hide_axes2=True)
 
-                x, y, sy, sx = model.data
+                x, y, sy, sx = self.datahandler.separated_data
                 kargs_errorbar["ecolor"] = symbol_color
                 kargs_scatter["c"] = symbol_color
                 # Making Plots
@@ -545,9 +530,7 @@ class SinglePlot(QObject):
                     sigma_x is False and sigma_y is False
                 ):  # Caso desconsiderar as duas
                     self.canvas.plot_scatter(x, y, kargs_scatter)
-                elif (
-                    sigma_x is False and sigma_y is True
-                ):  # Caso considerar só sy
+                elif sigma_x is False and sigma_y is True:  # Caso considerar só sy
                     sx = np.array([0] * len(x))
                     self.plot_data(x, y, sy, sx, kargs_errorbar, kargs_scatter)
                 else:  # Caso considerar só sx
@@ -586,6 +569,7 @@ class SinglePlot(QObject):
     def new(self):
         # Reseting canvas and model
         self.model.reset()
+        self.datahandler.reset()
         # self.canvas.reset()
         self.canvas.clear_axis()
         self.canvas.switch_axes(True)
@@ -624,19 +608,26 @@ class SinglePlot(QObject):
                 )
                 return 0
             # Loading data from the project
-            self.model.load_data(df_array=props["data"])
+            self.datahandler.load_data(df_array=props["data"])
+            self.model.data = self.datahandler.data
+            self.model._has_sx = self.datahandler.has_sx
+            self.model._has_sy = self.datahandler.has_sy
+
         else:
             try:
                 self.msg.raise_warn(
                     "O carregamento de arquivos antigos está limitado à uma versão anterior. Adaptação feita automaticamente."
                 )
                 props = self.load_old_json(props)
-            except:
+            except Exception as error:
                 self.msg.raise_error(
-                    "O arquivo carregado é incompatível com o ATUS."
+                    f"O arquivo carregado é incompatível com o ATUS.\n{error}"
                 )
                 return 0
-            self.model.load_data(df=props["data"])
+            self.datahandler.load_data(df_array=props["data"])
+            self.model.data = self.datahandler.data
+            self.model._has_sx = self.datahandler.has_sx
+            self.model._has_sy = self.datahandler.has_sy
 
         self.fill_plot_page(props)
 
@@ -679,13 +670,13 @@ class SinglePlot(QObject):
     def make_float(self, var, value=0.0):
         try:
             return float(var)
-        except:
+        except Exception:
             return value
 
     def make_int(self, var, value=0):
         try:
             return int(var)
-        except:
+        except Exception:
             return value
 
     @pyqtSlot(QJsonValue, result=int)
@@ -752,17 +743,17 @@ class SinglePlot(QObject):
                     "Nível de confiança deve ser sempre maior que zero e menor que 1. Rever nível de confiança."
                 )
                 return None
-        except:
+        except ValueError:
             pass
         try:
             ngl = ngl.replace(",", ".")
             ngl = float(ngl)
-        except:
+        except ValueError:
             pass
         try:
             mean = mean.replace(",", ".")
             mean = float(mean)
-        except:
+        except ValueError:
             pass
         try:
             std = std.replace(",", ".")
@@ -772,7 +763,7 @@ class SinglePlot(QObject):
                     "Desvio padrão deve ser sempre maior que zero. Rever desvio padrão."
                 )
                 return None
-        except:
+        except ValueError:
             pass
 
         s, x, y, x_area, y_area, title, xlabel, ylabel = interpreter_calculator(
