@@ -23,68 +23,68 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
-# from pytexit import py2tex
-# from PyQt5.QtCore import QObject, pyqtSlot
-# from io import BytesIO
-# import matplotlib.pyplot as plt
+from pytexit import py2tex
+from PyQt5.QtCore import QObject, pyqtSlot
+from io import BytesIO
+import matplotlib.pyplot as plt
 
 
-# class PyLatex(QObject):
-#     def __init__(self):
-#         super().__init__()
+class PyLatex(QObject):
+    def __init__(self):
+        super().__init__()
 
-#         self.fig = plt.figure(figsize=(0.03, 0.03))
-#         self.fig.text(
-#             x=0,
-#             y=0,
-#             s=f"INIT",
-#             fontsize=15,
-#             c="#FFF",
-#         )
-#         output = BytesIO()
-#         self.fig.savefig(
-#             output,
-#             dpi=500,
-#             transparent=True,
-#             format="svg",
-#             bbox_inches="tight",
-#             pad_inches=1,
-#         )
+        self.fig = plt.figure(figsize=(0.03, 0.03))
+        self.fig.text(
+            x=0,
+            y=0,
+            s="INIT",
+            fontsize=15,
+            c="#FFF",
+        )
+        output = BytesIO()
+        self.fig.savefig(
+            output,
+            dpi=500,
+            transparent=True,
+            format="svg",
+            bbox_inches="tight",
+            pad_inches=1,
+        )
 
-#     @pyqtSlot(str, result=str)
-#     def py2svg(self, formula, fontsize=15, dpi=500):
-#         """Render TeX formula to SVG.
-#         Args:
-#             formula (str): TeX formula.
-#             fontsize (int, optional): Font size.
-#             dpi (int, optional): DPI.
-#         Returns:
-#             str: SVG render.
-#         """
-#         try:
-#             svg = py2tex(formula, print_formula=False, print_latex=False)[1:-1]
-#         except:
-#             return ""
+    @pyqtSlot(str, result=str)
+    def py2svg(self, formula, fontsize=15, dpi=500):
+        """Render TeX formula to SVG.
+        Args:
+            formula (str): TeX formula.
+            fontsize (int, optional): Font size.
+            dpi (int, optional): DPI.
+        Returns:
+            str: SVG render.
+        """
+        try:
+            svg = py2tex(formula, print_formula=False, print_latex=False)[1:-1]
+        except Exception:
+            return ""
 
-#         self.fig.clear()
-#         self.fig.text(
-#             x=0,
-#             y=0,
-#             s=f"{svg}",
-#             fontsize=fontsize,
-#             c="#FFF",
-#         )
+        self.fig.clear()
+        self.fig.text(
+            x=0,
+            y=0,
+            s=f"{svg}",
+            fontsize=fontsize,
+            c="#FFF",
+        )
 
-#         output = BytesIO()
-#         self.fig.savefig(
-#             output,
-#             dpi=dpi,
-#             transparent=True,
-#             format="svg",
-#             bbox_inches="tight",
-#             pad_inches=1,
-#         )
-#         plt.close(self.fig)
+        output = BytesIO()
+        self.fig.savefig(
+            output,
+            dpi=dpi,
+            transparent=True,
+            format="svg",
+            bbox_inches="tight",
+            pad_inches=1,
+        )
+        plt.close(self.fig)
 
-#         output.seek(0)
-#         return output.read().decode("utf-8")
+        output.seek(0)
+        return output.read().decode("utf-8")
